backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
pandas==2.1.3
reportlab==4.0.7
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0

